Shooter autonomous odometry update and some shooter changes
